write businness :)

+++ TestApplication/Domain/Entities/Currency.cs
@@ -5,7 +5,7 @@ namespace Domain.Entities
 	public class Currency : BaseEntity
 	{
 		public string Name { get; set; }
-		public double RiskFreeRate { get; set; }
+		public byte RiskFreeRate { get; set; }
 
 		public ICollection<Company> Companies { get; set; }
 	}
